Changes to be committed: 	modified:   README.md 	modified:   scripts/CGRALib/cgra++ 	modified:   scripts/CGRALib/cgra.c 	modified:   scripts/CGRALib/cgra.cpp 	deleted:    scripts/CGRALib/cgra_Arkan.c 	deleted:    scripts/CGRALib/cgra_Vinh.c 	modified:   scripts/CGRALib/cgracc 	deleted:    scripts/CGRALib/diff.out
